agent_prospecteur/requirements.txt
requests
beautifulsoup4
openai
python-dotenv
googlemaps
sendgrid
playwright
pytest
